requirements.txt
certifi==2020.12.5
cycler==0.10.0
future==0.18.2
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pointnet2==0.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
six==1.15.0
torch==1.6.0
tqdm==4.60.0
trimesh==3.9.13
